+++ ClientApp/src/app/areas/measurement/components/addMeasurementType/addMeasurementType.component.ts
@@ -8,7 +8,7 @@ import { IMeasurementService } from 'services/IMeasurementService';
 import { INotificationService } from 'services/INotificationService';
 import { IRouterService } from 'services/IRouterService';
 
-import { AddMeasurementTypeRequest } from 'models/request/addMeasurementTypeRequest';
+import { AddMeasurementTypeRequest } from 'models/request/measurement/addMeasurementTypeRequest';
 import PaymentTypeResponse from 'models/response/payments/paymentTypeResponse';
 import { IPaymentService } from 'services/IPaymentService';
