+++ ClientApp/src/app/areas/measurement/components/updateMeasurement/updateMeasurement.component.ts
@@ -13,7 +13,7 @@ import { IMeasurementService } from 'services/IMeasurementService';
 import { INotificationService } from 'services/INotificationService';
 import { IRouterService } from 'services/IRouterService';
 
-import { AddMeasurementRequest } from 'models/request/addMeasurementRequest';
+import { AddMeasurementRequest } from 'models/request/measurement/addMeasurementRequest';
 import MeasurementTypeResponse from 'models/response/measurements/measurementTypeResponse';
 
 @Component({
