+++ ClientApp/src/app/areas/payments/components/updatePaymentType/updatePaymentType.component.ts
@@ -11,7 +11,7 @@ import { INotificationService } from 'services/INotificationService';
 import { IPaymentService } from 'services/IPaymentService';
 import { IRouterService } from 'services/IRouterService';
 
-import { AddPaymentTypeRequest } from 'models/request/addPaymentTypeRequest';
+import { AddPaymentTypeRequest } from 'models/request/payment/addPaymentTypeRequest';
 
 @Component({
     templateUrl: 'updatePaymentType.template.pug'
