+++ ClientApp/src/app/core/models/request/measurement/addMeasurementRequest.ts
