+++ ClientApp/src/app/core/models/request/stats/measurementStatisticsFilter.ts
@@ -1,0 +1,4 @@
+export default class MeasurementStatisticsFilter {
+    year: number;
+    measurementTypeId: number;
+}
