+++ ClientApp/src/app/core/models/request/stats/paymentStatisticsFilter.ts
@@ -1,0 +1,4 @@
+export default class PaymentStatisticsFilter {
+    year: number;
+    paymentTypeId: number;
+}
