--- ClientApp/src/app/core/models/response/payments/paymentStatsResponse.ts
@@ -1,11 +1,0 @@
-export interface GetPaymentsStatisticsDataItem {
-    month: number;
-    year: number;
-    amount: number;
-}
-
-export interface GetPaymentsStatisticsResponse {
-    year: number;
-    paymentTypeId: number;
-    statisticsData: Array<GetPaymentsStatisticsDataItem>;
-}
