+++ ClientApp/src/app/core/models/response/stats/measurementStatsResponse.ts
@@ -1,0 +1,11 @@
+export interface GetMeasurementStatisticsDataItem {
+    month: number;
+    year: number;
+    diff?: number;
+}
+
+export interface GetMeasurementStatisticsResponse {
+    year: number;
+    measurementTypeId: number;
+    statisticsData: Array<GetMeasurementStatisticsDataItem>;
+}
