+++ ClientApp/src/app/core/models/response/stats/paymentStatsResponse.ts
@@ -1,0 +1,11 @@
+export interface GetPaymentsStatisticsDataItem {
+    month: number;
+    year: number;
+    amount: number;
+}
+
+export interface GetPaymentsStatisticsResponse {
+    year: number;
+    paymentTypeId: number;
+    statisticsData: Array<GetPaymentsStatisticsDataItem>;
+}
