--- ClientApp/src/app/core/models/statisticsFilter.ts
@@ -1,4 +1,0 @@
-export default class StatisticsFilter {
-    year: number;
-    paymentTypeId: number;
-}
