+++ ClientApp/src/app/core/services/contracts/IPaymentService.ts
@@ -1,8 +1,8 @@
 import { Observable } from 'rxjs';
 
 import PaymentsFilter from 'models/paymentsFilter';
-import { AddPaymentRequest } from 'models/request/addPaymentRequest';
-import { AddPaymentTypeRequest } from 'models/request/addPaymentTypeRequest';
+import { AddPaymentRequest } from 'models/request/payment/addPaymentRequest';
+import { AddPaymentTypeRequest } from 'models/request/payment/addPaymentTypeRequest';
 import CommandExecutionResult from 'models/response/commandExecutionResult';
 import PaymentResponse from 'models/response/payments/paymentResponse';
 import PaymentTypeResponse from 'models/response/payments/paymentTypeResponse';
