+++ ClientApp/src/app/core/services/contracts/backend/IStatisticsApi.backend.ts
@@ -1,11 +1,10 @@
 import { Observable } from 'rxjs';
 
-import { GetPaymentsStatisticsResponse } from 'models/response/payments/paymentStatsResponse';
+import MeasurementStatisticsFilter from 'models/request/stats/measurementStatisticsFilter';
+import PaymentStatisticsFilter from 'models/request/stats/paymentStatisticsFilter';
 import QueryExecutionResult from 'models/response/queryExecutionResult';
-import StatisticsFilter from 'models/statisticsFilter';
 
 abstract class IStatisticsApiBackendService {
-    abstract getPaymentStatistics(filter: StatisticsFilter): Observable<QueryExecutionResult<GetPaymentsStatisticsResponse>>;
 }
 
 export { IStatisticsApiBackendService };
