+++ ClientApp/src/app/core/services/implementations/backend/measurementApi.backend.service.ts
@@ -9,8 +9,8 @@ import { isNullOrUndefined } from 'util';
 import { IMeasurementApiBackendService } from 'services/backend/IMeasurementApi.backend';
 
 import MeasurementsFilter from 'models/measurementsFilter';
-import { AddMeasurementRequest } from 'models/request/addMeasurementRequest';
-import { AddMeasurementTypeRequest } from 'models/request/addMeasurementTypeRequest';
+import { AddMeasurementRequest } from 'models/request/measurement/addMeasurementRequest';
+import { AddMeasurementTypeRequest } from 'models/request/measurement/addMeasurementTypeRequest';
 import CommandExecutionResult from 'models/response/commandExecutionResult';
 import MeasurementResponse from 'models/response/measurements/measurementResponse';
 import MeasurementsResponse from 'models/response/measurements/measurementsResponse';
