+++ ClientApp/src/app/core/services/implementations/backend/statisticsApi.backend.service.ts
@@ -8,9 +8,9 @@ import { isNullOrUndefined } from 'util';
 
 import { IStatisticsApiBackendService } from 'services/backend/IStatisticsApi.backend';
 
-import { GetPaymentsStatisticsDataItem, GetPaymentsStatisticsResponse } from 'models/response/payments/paymentStatsResponse';
+import MeasurementStatisticsFilter from 'models/request/stats/measurementStatisticsFilter';
+import PaymentStatisticsFilter from 'models/request/stats/paymentStatisticsFilter';
 import QueryExecutionResult from 'models/response/queryExecutionResult';
-import StatisticsFilter from 'models/statisticsFilter';
 
 @Injectable()
 class StatisticsApiBackendService implements IStatisticsApiBackendService {
@@ -23,7 +23,7 @@ class StatisticsApiBackendService implements IStatisticsApiBackendService {
     ) {
     }
 
-    public getPaymentStatistics(filter: StatisticsFilter): Observable<QueryExecutionResult<GetPaymentsStatisticsResponse>> {
+    public getPaymentStatistics(filter: PaymentStatisticsFilter): Observable<QueryExecutionResult<GetPaymentsStatisticsResponse>> {
         let params: HttpParams =
             new HttpParams();
