+++ ClientApp/src/app/core/services/implementations/payment.service.ts
@@ -9,8 +9,8 @@ import { IPaymentApiBackendService } from 'services/backend/IPaymentApi.backend'
 import { IPaymentService } from 'services/IPaymentService';
 
 import PaymentsFilter from 'models/paymentsFilter';
-import { AddPaymentRequest } from 'models/request/addPaymentRequest';
-import { AddPaymentTypeRequest } from 'models/request/addPaymentTypeRequest';
+import { AddPaymentRequest } from 'models/request/payment/addPaymentRequest';
+import { AddPaymentTypeRequest } from 'models/request/payment/addPaymentTypeRequest';
 import CommandExecutionResult from 'models/response/commandExecutionResult';
 import PaymentResponse from 'models/response/payments/paymentResponse';
 import PaymentTypeResponse from 'models/response/payments/paymentTypeResponse';
